source/backend/sensors_server/requirements.txt
requests
stomp.py